+++ src/sensors/FreescaleMPL115A2.cpp
@@ -20,7 +20,7 @@ FreescaleMPL115A2::FreescaleMPL115A2(TwoWire* theI2C, int8_t powerPin,
              MPL115A2_WARM_UP_TIME_MS, MPL115A2_STABILIZATION_TIME_MS,
              MPL115A2_MEASUREMENT_TIME_MS, powerPin, -1, measurementsToAverage,
              MPL115A2_INC_CALC_VARIABLES),
-      _i2c(theI2C) {}
+      _i2c(theI2C != nullptr ? theI2C : &Wire) {}
 // Delegating constructor
 FreescaleMPL115A2::FreescaleMPL115A2(int8_t  powerPin,
                                      uint8_t measurementsToAverage)