+++ src/sensors/PaleoTerraRedox.cpp
@@ -45,7 +45,7 @@ PaleoTerraRedox::PaleoTerraRedox(TwoWire* theI2C, int8_t powerPin,
              PTR_STABILIZATION_TIME_MS, PTR_MEASUREMENT_TIME_MS, powerPin, -1,
              measurementsToAverage),
       _i2cAddressHex(i2cAddressHex),
-      _i2c(theI2C) {}
+      _i2c(theI2C != nullptr ? theI2C : &Wire) {}
 PaleoTerraRedox::PaleoTerraRedox(int8_t powerPin, uint8_t i2cAddressHex,
                                  uint8_t measurementsToAverage)
     : Sensor("PaleoTerraRedox", PTR_NUM_VARIABLES, PTR_WARM_UP_TIME_MS,