+++ src/sensors/RainCounterI2C.cpp
@@ -40,7 +40,7 @@ RainCounterI2C::RainCounterI2C(TwoWire* theI2C, uint8_t i2cAddressHex,
              1),
       _rainPerTip(rainPerTip),
       _i2cAddressHex(i2cAddressHex),
-      _i2c(theI2C) {}
+      _i2c(theI2C != nullptr ? theI2C : &Wire) {}
 RainCounterI2C::RainCounterI2C(uint8_t i2cAddressHex, float rainPerTip)
     : Sensor("RainCounterI2C", BUCKET_NUM_VARIABLES, BUCKET_WARM_UP_TIME_MS,
              BUCKET_STABILIZATION_TIME_MS, BUCKET_MEASUREMENT_TIME_MS, -1, -1,