+++ ChangeLog.md
@@ -109,6 +109,8 @@ This affects the following defines:
     - `ANALOGELECCONDUCTIVITY_ADC_MAX_RATIO` (formerly a hardcoded value of `0.999f`)
 - **EverlightALSPT19**
   - Moved the calibration constant between current and lux to the `ALSPT19_UA_PER_1000LUX` preprocessor define.
+- **Bosch BMP3xx**
+  - Changed the default oversampling for both pressure and temperature to 1x (no oversampling) as recommended by the datasheet for weather and environmental monitoring.
 
 #### All Sensors
 
@@ -160,6 +162,13 @@ This is *not* breaking because only AVR and SAMD processors were supported anywa
 #### New Sensors
 
 - **NEW SENSOR** Added a new sensor for simple analog voltage using the built-in processor ADC
+- **NEW SENSOR** Added support for the TE Connectivity (Meas Specialties) MS5837
+  - This is the sensor embedded in the Blue Robotics Bar02 and Bar30 sensors.
+
+#### New Features for Specific Sensors
+
+- **TIADS1x15**
+  - Added support for a secondary hardware I2C instance.
 
 #### Features for All Sensors